Refactor to use vLLM for LLM and embedding models
Migrates LLM and Qwen3-VL embedding models to use vLLM for improved performance and scalability. Updates Dockerfile and requirements to install vLLM and related dependencies, removes bitsandbytes, and adjusts model loading and inference logic accordingly. Moves PerceptualHasher to a new hashing module and updates imports throughout the codebase. Adds new embedding configuration options and updates default batch size.

+++ preprocessor/video/frame_subprocessors.py
@@ -21,7 +21,7 @@ from preprocessor.core.base_processor import (
     ProcessingItem,
 )
 from preprocessor.embeddings.gpu_batch_processor import GPUBatchProcessor
-from preprocessor.embeddings.image_hasher import PerceptualHasher
+from preprocessor.hashing.image_hasher import PerceptualHasher
 from preprocessor.utils.batch_processing_utils import (
     compute_embeddings_in_batches,
     compute_hashes_in_batches,
@@ -142,7 +142,7 @@ class VideoEmbeddingSubProcessor(FrameSubProcessor):
             console.print(f"[cyan]Loading embedding model: {self.model_name}[/cyan]")
             self.model = Qwen3VLEmbedder(
                 model_name_or_path=self.model_name,
-                dtype=torch.float16,
+                torch_dtype=torch.bfloat16,
             )
             self.gpu_processor = GPUBatchProcessor(
                 self.model,
